Fixing issues with ProfileClick
